pyproject.toml
[tool.poetry]
name = "aws-cdk-organizations-demo"
version = "0.0.1"
description = "DEMO of an AWS Organization built on top of cdk-organizations with best IaC standards"
authors = ["Santiago Garcia Arango <san99tiago@gmail.com>"]
license = "Apache"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
aws-cdk-lib = "2.93"
constructs = ">=10.0.0,<11.0.0"
pepperize-cdk-organizations = "^0.7.680"

[tool.poetry.group.dev.dependencies]
poethepoet = "^0.20.0"
black = "^23.9.1"
pytest = "^7.3.2"
pytest-mock = "^3.11.1"
coverage = "^7.2.7"
moto = "^4.1.11"
boto3 = "^1.26.153"

[tool.pytest.ini_options]
minversion = "6.0"
pythonpath = [
    "cdk",
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poe.tasks]
test-unit = ["_test_unit", "_coverage_html"]
black-format = "black ."
black-check = "black . --check --diff -v"
_test_unit = "coverage run -m pytest tests/unit"
_coverage_html = "coverage html"

[tool.coverage.run]
branch = true
source = ["cdk"]
omit = [
    "**/__init__.py"
]

[tool.coverage.report]
show_missing = false
